refactor : 에러 위치 수정
