+++ src/main/java/roomescape/dataLayer/errors/ReservationNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.model.errors;
+package roomescape.dataLayer.errors;
 
 public class ReservationNotFoundException extends IllegalArgumentException {
     public ReservationNotFoundException(String s) {
