+++ src/main/java/roomescape/dataLayer/errors/TimeNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.model.errors;
+package roomescape.dataLayer.errors;
 
 public class TimeNotFoundException extends IllegalArgumentException {
     public TimeNotFoundException(String s) {
